Use ES2021 as output for ESNext breaks some recent browsers (#289)
Safari doesn;t support private class fields natively until 14.4. This means that anyone on iOS on a version less than that will have a broken experience. This changes the target output to use ES2021 this should transpile the minimum amount of features needed to have this still work in modern-ish browsers. In the interests of keeping Web3 accessible to folk who dont have the latest devices of OS's I think we should introduce a small amount of transpilation for a bigger support.